pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wsi-tissue-pipeline"
version = "0.1.0"
description = "Reproducible whole-slide image processing pipeline for tissue section extraction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dominic Padova", email = "dpadova95@gmail.com"}
]
keywords = [
    "whole-slide-imaging",
    "microscopy",
    "ome-zarr",
    "tissue-segmentation",
    "medical-imaging",
    "neuroglancer"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "dask[array,distributed]>=2024.1.0",
    "dask-image>=2024.1.0",
    
    # Image processing
    "scikit-image>=0.21.0",
    "Pillow>=10.0.0",
    "imageio>=2.31.0",
    "tifffile>=2024.1.30",
    "opencv-python-headless>=4.8.0",
    "tinybrain>=1.3.0",
    
    # OME-Zarr and storage
    "zarr>=2.16.0",
    "ngff-zarr>=0.4.0",
    "tensorstore>=0.1.45",
    "numcodecs>=0.12.0",
    
    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
    
    # Utilities
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

mlflow = [
    "mlflow>=2.10.0",
]

visualization = [
    "neuroglancer>=2.37",
    "matplotlib>=3.7.0",
]

cloudvolume = [
    "cloud-volume>=8.0.0",
]

torch = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

pathology = [
    # Optional pathology-specific tools
    # "tiatoolbox>=1.5.0",  # Uncomment if needed
    # "pathml>=2.0.0",       # Uncomment if needed
]

all = [
    "wsi-tissue-pipeline[dev,mlflow,visualization,torch,cloudvolume]",
]

[project.scripts]
wsi-pipeline = "wsi_pipeline.cli:main"

[project.urls]
Homepage = "https://github.com/dmp2/wsi-tissue-pipeline"
Documentation = "https://github.com/dmp2/wsi-tissue-pipeline#readme"
Repository = "https://github.com/dmp2/wsi-tissue-pipeline.git"
Issues = "https://github.com/dmp2/wsi-tissue-pipeline/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wsi_pipeline = ["*.yaml", "*.json"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=wsi_pipeline --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/wsi_pipeline"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
